+++ server/cache/index.js
@@ -1,7 +1,7 @@
 require('dotenv').config()
 const { pool } = require('./database')
-const { users, findUser, createUser, initializeUser, updatePassword, refreshTokens , updateUserImage,getSuppliersCategories_Name } = require('./user')
-const { products, initializeProduct, findProduce } = require('./product')
+const { users, findUser, createUser, initializeUser, updatePassword, refreshTokens , updateUserImage } = require('./user')
+const { products, initializeProduct, findProduce , getSuppliersCategories_Name} = require('./product')
 const { categories, initializeCategory } = require('./category')
 const { orders, initializeOrder } = require('./order')
 const { suppliers, initializeSupplier } = require('./supplier')
