+++ server/cache/product.js
@@ -53,8 +53,24 @@ async function findProduce(name) {
     }
     return product
 }
+
+async function getSuppliersCategories_Name() {
+    const queryString1 = 'select supplier_name from suppliers;';
+    const queryString2 = 'select category_name  from categories;';
+
+    try {
+        const [rows1] = await pool.query(queryString1)
+        const [rows2] = await pool.query(queryString2)
+        return {rows1 , rows2}
+    } catch (error) {
+        sole.log('\x1b[31m%s\x1b[0m', error.message)
+        throw new Error(error.message)
+    }
+}
+
 module.exports = {
     initializeProduct,
     findProduce,
+    getSuppliersCategories_Name,
     products
 }
