+++ server/routes/product.js
@@ -4,5 +4,6 @@ const { productController } = require('../controllers')
 
 router.get('/product', productController.getProductByID)
 router.get("/search",productController.searchProduct)
+router.get("/suppliers-categories", productController.getSuppliers_categories)
 
 module.exports = router
